Output files for Molex to check
1. Flex material (polyimide) thickness (s)? see Fab drawing (0.012mm
polyimide)
If high powered LED’s, copper foil thickness?   0.5 Oz. copper (Bench
mark by PCB trace width calculator)
Is this intended to be a rigidized flex or a rigid flex? rigidized
flex, just on the bottom surface
Note: this would be decided by if components (LED’s) are being mounted
to just top surface, or bottom surface as well.
Is quote for unassembled flex or for flex assembly? Unassembled flex
If assembly, we need BOM.
If not assembly, what is required surface finish? ENIG
If not assembly, are part being delivered as singulated or as an array?
singulated
If array, what are the details?